ai-agents/requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.0
pydantic-settings>=2.0
httpx>=0.27.0
asyncpg>=0.30.0
redis>=5.0
Pillow>=11.0
structlog>=24.0
pytest>=8.0
pytest-asyncio>=0.24.0
